Update GFDL_PPAN_instructions.md
I learned recently that a cleaner and more robust way of installing `numpy` without MKL optimizations is to specify `nomkl` as a dependency when installing via `conda`.  See [here](https://docs.anaconda.com/mkl-optimizations/#uninstalling-mkl) for reference.  This has the added benefit that it allows use of the most recent `numpy` version too.